Move node resolver into separate file
