+++ cmd/frontend/graphqlbackend/graphqlbackend.go
@@ -370,211 +370,6 @@ func NewSchema(db dbutil.DB, batchChanges BatchChangesResolver, codeIntel CodeIn
 	)
 }
 
-type Node interface {
-	ID() graphql.ID
-}
-
-type NodeResolver struct {
-	Node
-}
-
-func (r *NodeResolver) ToAccessToken() (*accessTokenResolver, bool) {
-	n, ok := r.Node.(*accessTokenResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToMonitor() (MonitorResolver, bool) {
-	n, ok := r.Node.(MonitorResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToMonitorQuery() (MonitorQueryResolver, bool) {
-	n, ok := r.Node.(MonitorQueryResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToMonitorEmail() (MonitorEmailResolver, bool) {
-	n, ok := r.Node.(MonitorEmailResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToMonitorActionEvent() (MonitorActionEventResolver, bool) {
-	n, ok := r.Node.(MonitorActionEventResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToMonitorTriggerEvent() (MonitorTriggerEventResolver, bool) {
-	n, ok := r.Node.(MonitorTriggerEventResolver)
-	return n, ok
-}
-
-// TODO(campaigns-deprecation): This should be removed once we remove campaigns completely
-func (r *NodeResolver) ToCampaign() (BatchChangeResolver, bool) {
-	if n, ok := r.Node.(BatchChangeResolver); ok {
-		return n, n.ActAsCampaign()
-	}
-	return nil, false
-}
-
-// TODO(campaigns-deprecation): This should be removed once we remove campaigns completely
-func (r *NodeResolver) ToCampaignSpec() (BatchSpecResolver, bool) {
-	if n, ok := r.Node.(BatchSpecResolver); ok {
-		return n, n.ActAsCampaignSpec()
-	}
-	return nil, false
-}
-
-func (r *NodeResolver) ToBatchChange() (BatchChangeResolver, bool) {
-	if n, ok := r.Node.(BatchChangeResolver); ok {
-		// TODO(campaigns-deprecation): This should be removed once we remove campaigns completely
-		return n, !n.ActAsCampaign()
-	}
-	return nil, false
-}
-
-func (r *NodeResolver) ToBatchSpec() (BatchSpecResolver, bool) {
-	if n, ok := r.Node.(BatchSpecResolver); ok {
-		// TODO(campaigns-deprecation): This should be removed once we remove campaigns completely
-		return n, !n.ActAsCampaignSpec()
-	}
-	return nil, false
-}
-
-func (r *NodeResolver) ToExternalChangeset() (ExternalChangesetResolver, bool) {
-	n, ok := r.Node.(ChangesetResolver)
-	if !ok {
-		return nil, false
-	}
-	return n.ToExternalChangeset()
-}
-
-func (r *NodeResolver) ToHiddenExternalChangeset() (HiddenExternalChangesetResolver, bool) {
-	n, ok := r.Node.(ChangesetResolver)
-	if !ok {
-		return nil, false
-	}
-	return n.ToHiddenExternalChangeset()
-}
-
-func (r *NodeResolver) ToChangesetEvent() (ChangesetEventResolver, bool) {
-	n, ok := r.Node.(ChangesetEventResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToHiddenChangesetSpec() (HiddenChangesetSpecResolver, bool) {
-	n, ok := r.Node.(ChangesetSpecResolver)
-	if !ok {
-		return nil, ok
-	}
-	return n.ToHiddenChangesetSpec()
-}
-
-func (r *NodeResolver) ToVisibleChangesetSpec() (VisibleChangesetSpecResolver, bool) {
-	n, ok := r.Node.(ChangesetSpecResolver)
-	if !ok {
-		return nil, ok
-	}
-	return n.ToVisibleChangesetSpec()
-}
-
-// TODO(campaigns-deprecation): This should be removed once we remove campaigns completely
-func (r *NodeResolver) ToCampaignsCredential() (CampaignsCredentialResolver, bool) {
-	n, ok := r.Node.(CampaignsCredentialResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToBatchChangesCredential() (BatchChangesCredentialResolver, bool) {
-	n, ok := r.Node.(BatchChangesCredentialResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToProductLicense() (ProductLicense, bool) {
-	n, ok := r.Node.(ProductLicense)
-	return n, ok
-}
-
-func (r *NodeResolver) ToProductSubscription() (ProductSubscription, bool) {
-	n, ok := r.Node.(ProductSubscription)
-	return n, ok
-}
-
-func (r *NodeResolver) ToExternalAccount() (*externalAccountResolver, bool) {
-	n, ok := r.Node.(*externalAccountResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToExternalService() (*externalServiceResolver, bool) {
-	n, ok := r.Node.(*externalServiceResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToGitRef() (*GitRefResolver, bool) {
-	n, ok := r.Node.(*GitRefResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToRepository() (*RepositoryResolver, bool) {
-	n, ok := r.Node.(*RepositoryResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToUser() (*UserResolver, bool) {
-	n, ok := r.Node.(*UserResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToOrg() (*OrgResolver, bool) {
-	n, ok := r.Node.(*OrgResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToOrganizationInvitation() (*organizationInvitationResolver, bool) {
-	n, ok := r.Node.(*organizationInvitationResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToGitCommit() (*GitCommitResolver, bool) {
-	n, ok := r.Node.(*GitCommitResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToRegistryExtension() (RegistryExtension, bool) {
-	if NodeToRegistryExtension == nil {
-		return nil, false
-	}
-	return NodeToRegistryExtension(r.Node)
-}
-
-func (r *NodeResolver) ToSavedSearch() (*savedSearchResolver, bool) {
-	n, ok := r.Node.(*savedSearchResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToSearchContext() (*searchContextResolver, bool) {
-	n, ok := r.Node.(*searchContextResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToSite() (*siteResolver, bool) {
-	n, ok := r.Node.(*siteResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToLSIFUpload() (LSIFUploadResolver, bool) {
-	n, ok := r.Node.(LSIFUploadResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToLSIFIndex() (LSIFIndexResolver, bool) {
-	n, ok := r.Node.(LSIFIndexResolver)
-	return n, ok
-}
-
-func (r *NodeResolver) ToOutOfBandMigration() (*outOfBandMigrationResolver, bool) {
-	n, ok := r.Node.(*outOfBandMigrationResolver)
-	return n, ok
-}
-
 // schemaResolver handles all GraphQL queries for Sourcegraph. To do this, it
 // uses subresolvers which are globals. Enterprise-only resolvers are assigned
 // to a field of EnterpriseResolvers.
@@ -629,84 +424,6 @@ var EnterpriseResolvers = struct {
 // DEPRECATED
 func (r *schemaResolver) Root() *schemaResolver {
 	return &schemaResolver{db: r.db}
-}
-
-func (r *schemaResolver) Node(ctx context.Context, args *struct{ ID graphql.ID }) (*NodeResolver, error) {
-	n, err := r.nodeByID(ctx, args.ID)
-	if err != nil {
-		return nil, err
-	}
-	if n == nil {
-		return nil, nil
-	}
-	return &NodeResolver{n}, nil
-}
-
-func (r *schemaResolver) nodeByID(ctx context.Context, id graphql.ID) (Node, error) {
-	switch relay.UnmarshalKind(id) {
-	case "AccessToken":
-		return accessTokenByID(ctx, r.db, id)
-	case "Campaign":
-		return r.CampaignByID(ctx, id)
-	case "BatchChange":
-		return r.BatchChangeByID(ctx, id)
-	case "CampaignSpec":
-		return r.CampaignSpecByID(ctx, id)
-	case "BatchSpec":
-		return r.BatchSpecByID(ctx, id)
-	case "ChangesetSpec":
-		return r.ChangesetSpecByID(ctx, id)
-	case "Changeset":
-		return r.ChangesetByID(ctx, id)
-	case "CampaignsCredential":
-		return r.CampaignsCredentialByID(ctx, id)
-	case "BatchChangesCredential":
-		return r.BatchChangesCredentialByID(ctx, id)
-	case "ProductLicense":
-		if f := ProductLicenseByID; f != nil {
-			return f(ctx, r.db, id)
-		}
-		return nil, errors.New("not implemented")
-	case "ProductSubscription":
-		if f := ProductSubscriptionByID; f != nil {
-			return f(ctx, r.db, id)
-		}
-		return nil, errors.New("not implemented")
-	case "ExternalAccount":
-		return externalAccountByID(ctx, r.db, id)
-	case externalServiceIDKind:
-		return externalServiceByID(ctx, r.db, id)
-	case "GitRef":
-		return r.gitRefByID(ctx, id)
-	case "Repository":
-		return r.repositoryByID(ctx, id)
-	case "User":
-		return UserByID(ctx, r.db, id)
-	case "Org":
-		return OrgByID(ctx, r.db, id)
-	case "OrganizationInvitation":
-		return orgInvitationByID(ctx, r.db, id)
-	case "GitCommit":
-		return r.gitCommitByID(ctx, id)
-	case "RegistryExtension":
-		return RegistryExtensionByID(ctx, r.db, id)
-	case "SavedSearch":
-		return r.savedSearchByID(ctx, id)
-	case "Site":
-		return r.siteByGQLID(ctx, id)
-	case "LSIFUpload":
-		return r.LSIFUploadByID(ctx, id)
-	case "LSIFIndex":
-		return r.LSIFIndexByID(ctx, id)
-	case "CodeMonitor":
-		return r.MonitorByID(ctx, id)
-	case "OutOfBandMigration":
-		return r.OutOfBandMigrationByID(ctx, id)
-	case "SearchContext":
-		return r.SearchContextByID(ctx, id)
-	default:
-		return nil, errors.New("invalid id")
-	}
 }
 
 func (r *schemaResolver) Repository(ctx context.Context, args *struct {
